Allow unplaced seqs in reference, take 1
Previously, benchmarking relied on each reference sequence being placed
unambiguously in a single genome, and a single location within that genome.
When dealing with real assemblies, the source of a given contig must be found
by mapping contigs back to references, which create some ambiguity.

In this commit, sequences can belong to a position in a genome, a genome, or
a clade in the reference.
Sequences only count towards recall if they are mapped to a position, but count
towards precision for a given clade if they belong to that clade or a child of
that clade.

However, this commit makes the benchmark extremely strict, to the point where
it's nearly impossible to reconstruct any genome at all with real assemblies.
Hence I suspect I will need to rethink and rehaul benchmarking again soon.