Allow ambiguously mapped contigs
The reference can now be "ambiguous", in that sequences can map to several
locations, each being a (source::String, from::Int, to::Int). When counting
true positivies, all mapping locations is considered recovered. For false
positives, this is the sum of lengths of all contigs not mapping to the genome.
False negatives is the genome size minus false positives.