Store both recovered assemblies and genomes
This commit once again significantly rehauls the benchmarking process.
Now, we want to keep track of how large part of the genomes are assembled,
and then benchmark against both the total genome size and against the total
assembly size at the same time.
I.e, we want two benchmarks, where one considers how many genomes are recovered
at various recall/precision levels, and one which considers how many of the
assemblies present in the reference are recovered. The latter is much easier.

So, with this commit, we can compare
* Quality of the assembly relative to the genomes
* Quality of binning relative to genomes
* Quality of binning relative to assembly