Add version field to ReferenceJSON
I've been making a lot of changes to ReferenceJSON files, each of them breaking.
In order to be able to develop VambBenchmarks further, there needs to be some
kind of mechanism to figure out what version of VambBenchmarks goes with a given
outdated reference file.
Now, users can check the git log, or use `git blame` to figure out which version
of VambBenchmarks they should use for a given reference created from now on.