Do not count ghost sequences as contamination
Previously, if a sequence was not mapping to any source, it was considered
"foreign" to all genomes.
Now, only count a sequence as foreign if it has a mapping, but not to the target
genome.

Here is the difference for the same clustering of the pTracker CAMI2 Oral data
assembled with SPAdes:

Before:
P\R   0.3  0.4  0.5  0.6  0.7  0.8  0.9 0.95 0.99
0.5   330  310  296  271  221  140   61   40   14
0.6   285  271  260  238  195  121   56   38   13
0.7   247  232  225  208  174  107   48   34   11
0.8   190  178  170  158  126   77   40   32   10
0.9   136  120  117  111   89   57   37   31   10
0.95   95   80   76   73   65   49   35   30   10
0.99   49   44   43   41   38   37   31   27   10

After:
P\R   0.3  0.4  0.5  0.6  0.7  0.8  0.9 0.95 0.99
0.5   348  327  313  287  236  152   67   41   15
0.6   308  288  275  251  207  130   61   38   13
0.7   275  258  246  225  189  118   54   35   12
0.8   219  205  196  179  150   92   47   34   12
0.9   158  141  138  129  107   71   44   33   12
0.95  123  107  103   98   79   58   41   33   12
0.99   77   66   64   60   53   51   37   32   12