Add kwarg filter_genomes to Binning
This allows passing in a function to decide to ignore some genomes. We found
that it is insufficient to filter the genomes away in the Reference, because
the removal of the genomes from the reference means that any sequences mapping
to those genomes will now map to nothing, meaning they no longer count as
contamination.

Sometimes though, we want to keep the information that some sequence map to e.g.
this virus, and yet ignore all viruses when benchmarking.