Simplify taxmap structure slightly
Now, do not allow parents in the taxmap to be `nothing`. The invariants expected
in `parse_taxonomy` is stated more clearly, and the format is less complex.
This is a breaking change, although it's not expected to break the existing
reference file in practise.

Also, this PR extracts the types of the fields of `ReferenceJSON` to their own
types. For now, this is used to enforce correct typing throughout this package's
internals, but in the future, these types can be used when creating references.