Remove min_seq_length kwarg from Reference
We shouldn't make it too easy to filter the reference - instead, the Binning
should be filtered.
There are two reasons for that: One: Usually people want to filter the Binning,
see the comment to commit 78d3ed3, and two: Removing contigfs from the ref means
that these cannot be present in the Binning.