Support alignments to circular subjects
Most microbial genomes are circular. Hence, when they are assembled and stored
as linear strings, the start and ending points are arbitrary, and we must expect
contigs to align to these subjects from the end of the subject to the beginning.
I.e. if subject S is a circular sequence represented as a linear string, and C
is a contig from S, then we must expect that C can align to S like this:

S  =============================
C  ---->                  |---->
       stop               start

We now support this by representing the span where the sequences align to, not
with `UnitRange`, but with `Tuple{Int, Int}`, representing `(start, stop)`. Both
of these number must be in `1:length`, but now, we allow `start > stop`.

This somewhat complicates computing coverage, which is a performance bottleneck
of the program, but we need to support this. If necessary, we can optimise this
later.