Remove recoverable_genomes field from Binning
The value in that field was computed based on the number of assembled genomes
in the Binning's Reference. However, we do not guarantee that all sequences in
the Reference are present in the Binning, so the number could be misleading.
Computing the true number is more involved and requires creating a gold standard
Binning based on the sequences present in the given Binning.