Switch default from assemblies to genomes
Before, BBB defaulted to benchmarking the number of reconstructed assemblies,
not genomes. This was motivated by the fact that genomes are harder than asms,
because many genomes are lost in the assembly step already.
We wanted a higher baseline number when binning, such that we could more easily
track improvements to binning.

However, asms are less biologically important than genomes, are more susceptible
to spurious performance measures, since some asms are only a few kbp large.
Reconstructing these are trivial, and unimportant. So, I've decided to switch
the default to genomes.