Git commit doc Manifest.toml
Normally, the Manifest should not be trakced by git, since it contains the full
resolved dependency graph, which users are neither interested in, nor want to
be contrained by.
However, for docs, it might be important for users in the future to be able to
get an exact environment to build the BBB docs in. Also, it doesn't matter if
users are constrained to a specific environment when building the docs.