Bugfix: Filter genome on Bin construction
The precision/recall values between Genome/Bin and Clade/Bin pairs are computed
when the Bin objects are instantiated shortly after parsing the clusters, but
before calling the `benchmark` function, whose job it is to compute the summary
statistics stored in the Binning object itself.

Before this commit, passing `filter_genomes` to a Binning would instantiate the
Bins without filtering, and then filter them when computing the summary stats.
This was done by simply skipping the removed genomes during the computation of
summary statistics.
However, the filtering of genomes did not affect statistics for Clades, because
clades could not simply be skipped (since only _some_ of their genomes might
be removed). This resulted in wrong (unfiltered) clade statistics.

In this commit, filter the genomes already when the Bins are constructed.