Also report n bins at asm recall
BBB is counts both the number of bins that correspond to a genome at a given
recall/precision threshold, and also the number of genomes recovered at a r/p
threshold.
Further we can count recall=1.0 as being either the full assembly, or the full
genome sequence.
This gives 2x2 = 4 possibilities. Before this commit, BBB only computed three
of these four possibilities, missing the bin/asm combination.

Add this combination, and do some slight refactoring of the benchmark function.
That probably should be completely refactored at this point, but it's tricky.