Allow constructing a Binning from explicit Bins (#48)
Document and test another method of Binning, which creates a Binning from a vec
of Bin. This allows users to manually construct their bins instead of having to
load them from a file.

Similar, add a new `filter_bins` kwarg to Binning, which removes any bin not
passing the predicate.