Make Binning construction more type stable
One of the main methods of Binning uses functions as keyword arguments. Since
kwargs don't participate in dispatch, Julia does not specialize on kwargs.
This means that the functions are inferred to just Function, such that calling
them returns Any.
Sidestep this by using the function barrier trick where the functions are used.