Migrate to JSON.jl v1 (#64)
JSON.jl has a new v1 release with a newer, better API. Use it.
This has two reasons: One is simply that JSON.jl will be maintained whereas
the old JSON3 probably will not be.
The second is that we need some of the new API to partially parse the references,
which is worked around with a regex hack previously.

This is currently a slight performance regression. However, I believe it's worth it.